Add LogMgr and LogIterator

+++ src/file/blockid.rs
@@ -1,10 +1,11 @@
+#[derive(Clone)]
 pub struct BlockId {
     filename: String,
-    blknum: usize,
+    blknum: i32,
 }
 
 impl BlockId {
-    pub fn new(filename: &str, blknum: usize) -> BlockId {
+    pub fn new(filename: &str, blknum: i32) -> BlockId {
         BlockId {
             filename: filename.to_string(),
             blknum,
@@ -15,7 +16,7 @@ impl BlockId {
         &self.filename
     }
 
-    pub fn number(&self) -> usize {
+    pub fn number(&self) -> i32 {
         self.blknum
     }
 }
